Rename: 모달스토어 이름 변경
